+++ src/BabelFish/DataActors/Tournaments/AverageMethodConfiguration.cs
@@ -1,10 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
 namespace Scopos.BabelFish.DataActors.Tournaments {
+
+    /// <summary>
+    /// Concrete MergeConfiguration class for the AverageMethod result list merger.
+    /// </summary>
     public class AverageMethodConfiguration : MergeConfiguration {
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
         public AverageMethodConfiguration() : base() {
 
             /*
@@ -24,5 +27,11 @@ namespace Scopos.BabelFish.DataActors.Tournaments {
         /// If true, scores with a DNF are excluded when calculating a participant's average.
         /// </summary>
         public bool ExcludeDNFFromAverage { get; set; } = false;
+
+        /// <summary>
+        /// Parameter that sets how many of the top scores to count towards a participant's average.
+        /// <para>A value of 0 (the default) means to count all scores.</para>
+        /// </summary>
+        public int CountTopScores { get; set; } = 0;
     }
 }