+++ src/BabelFish/DataModel/OrionMatch/Squadding.cs
@@ -1,21 +1,27 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Scopos.BabelFish.DataModel;
-
 namespace Scopos.BabelFish.DataModel.OrionMatch {
+
+    /// <summary>
+    /// Class representing a Participant and their squadding assignment in a Match's event. 
+    /// </summary>
     public class Squadding : IRLIFItem {
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
         public Squadding() { }
 
-		public Participant Participant { get; set; }
+        /// <summary>
+        /// The Participant (could be individual or team).
+        /// </summary>
+        public Participant Participant { get; set; }
 
-		public SquaddingAssignment SquaddingAssignment { get; set; }
+        /// <summary>
+        /// the Participant's SquaddingAssignment.
+        /// </summary>
+        public SquaddingAssignment SquaddingAssignment { get; set; }
 
-		public override string ToString() {
+        /// <inheritdoc />
+        public override string ToString() {
             return $"Squadding for {Participant.DisplayName}: {SquaddingAssignment}";
         }
     }